+++ src/init.lua
@@ -1,6 +1,7 @@
 local RoactSpring = {
 	useSpring = require(script.useSpring),
 	useSprings = require(script.useSprings),
+	useTrail = require(script.useTrail),
 	Controller = require(script.Controller),
 	SpringValue = require(script.SpringValue),
 	config = require(script.constants).config,