+++ stories/hooks/useSpringImmediate.story.lua
@@ -7,7 +7,7 @@ local CircleButton = require(script.Parent.Parent.components.CircleButton)
 
 local e = Roact.createElement
 
-local function Button(props, hooks)
+local function Button(_, hooks)
     local styles, api = RoactSpring.useSpring(hooks, function()
         return { position = UDim2.fromScale(0.5, 0.5), }
     end)